:bug: Fix: Fixed CustomAlert close by Enterkey and Seperated from styles
Fix: Fixed CustomAlert close by Enterkey and Seperated from styles
