+++ src/components/Common/Post/Post.jsx
@@ -20,7 +20,7 @@ import { uploadPost } from '../../../api/post';
 import { uploadImgs } from '../../../api/imageApi';
 import myProfile from '../../../assets/image/profile-img4.svg';
 import uploadChat from '../../../assets/image/chat-upload-btn.svg';
-import CustomAlert from '../../../pages/LoginPage/CustomAlert';
+import CustomAlert from '../CustomAlert/CustomAlert';
 
 export default function Post({
   id,
